# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medshelf API",
        "description": "OpenAPI documentation for the Medshelf API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base URL"
        }
    ],
    "paths": {
...
        "/profiles/{profileId}/treatments": {
            "get": {
                "tags": [
                    "Treatments"
                ],
                "summary": "List treatments for profile",
                "operationId": "1b6cb959e2e6dcfda08b24ab89a563b0",
                "parameters": [
                    {
                        "name": "profileId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "required": [
                                                "items",
                                                "nextCursor"
                                            ],
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TreatmentView"
                                                    }
                                                },
                                                "nextCursor": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "required": [
                                                "items",
                                                "totalCount",
                                                "page",
                                                "size",
                                                "hasMorePages"
                                            ],
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TreatmentView"
                                                    }
                                                },
                                                "totalCount": {
                                                    "type": "integer",
                                                    "minimum": 0
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "size": {
                                                    "type": "integer",
                                                    "maximum": 100,
                                                    "minimum": 1
                                                },
                                                "hasMorePages": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Create treatment",
                "operationId": "89e739e5464b4cc08c68f8952d1b6df9",
                "parameters": [
                    {
                        "name": "profileId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "itemId",
                                    "dose",
                                    "frequencyUnit",
                                    "startDate"
                                ],
                                "properties": {
                                    "itemId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "dose": {
                                        "type": "number",
                                        "minimum": 0.01,
                                        "example": 1.5
                                    },
                                    "frequencyUnit": {
                                        "type": "string",
                                        "enum": [
                                            "hours",
                                            "days",
                                            "weeks"
                                        ]
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-11"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TreatmentResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/treatments/{treatmentId}/qr": {
            "get": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Generate QR code image for a treatment",
                "operationId": "be6dff99f9e82416a0024fb81e39d286",
                "parameters": [
                    {
                        "name": "treatmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PNG image of the QR code encoding the treatment summary",
                        "content": {
                            "image/png": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "required": [
                    "message",
                    "timestamp"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "TreatmentView": {
                "required": [
                    "id",
                    "profile",
                    "item",
                    "status",
                    "dose",
                    "frequencyUnit",
                    "startDate"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "profile": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "item": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "product": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "status": {
                        "type": "string"
                    },
                    "dose": {
                        "type": "number"
                    },
                    "frequencyUnit": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}